+++ Sources/VCore/Models/CastingError.swift
@@ -9,28 +9,21 @@ import Foundation
 
 // MARK: - Casting Error
 /// Error that occurs during casting.
-public struct CastingError: VCoreError, Equatable, Sendable {
-    // MARK: Properties
-    private let fromType: String
-    private let toType: String
-
+@MemberwiseInitializable(accessLevelModifier: .private)
+public struct CastingError: BaseErrorProtocol, Sendable {
+    public static let domain: String = "com.vcore"
+    public let code: Int
+    public let description: String
+    
     // MARK: Initializers
-    /// Initializes `CastingError` with types.
+    /// Indicates that casting has failed.
     public init(
         from fromType: String,
         to toType: String
     ) {
-        self.fromType = fromType
-        self.toType = toType
-    }
-
-    // MARK: Properties - VCore Error
-    public let vCoreErrorCode: Int = 0
-
-    public var vCoreErrorDescription: String { "Failed to cast '\(fromType)' to '\(toType)'" }
-
-    // MARK: Equatable
-    public static func == (lhs: Self, rhs: Self) -> Bool {
-        isEqual(lhs, to: rhs, by: \.fromType, \.toType)
+        self.init(
+            code: 0,
+            description: "Failed to cast '\(fromType)' to '\(toType)'"
+        )
     }
 }